Move types from mock to global types
